feat: add client home page component
