+++ src/components/home/index.ts
@@ -7,3 +7,4 @@ export { default as QuickActions } from "./QuickActions";
 export { default as TeamPromptsCard } from "./TeamPromptsCard";
 export { default as BottomNav } from "./BottomNav";
 export { default as IsometricRoom } from "./IsometricRoom";
+export { default as HomePage } from "./HomePage";
